[Core] Consistent moving DelaunatorUtilities to tessellation_utilities folder
