+++ kratos/modeler/cad_tessellation_modeler.cpp
@@ -13,7 +13,7 @@
 
 // Project includes
 #include "includes/define.h"
-#include "utilities/delaunator_utilities.h"
+#include "utilities/tessellation_utilities/delaunator_utilities.h"
 #include "utilities/parallel_utilities.h"
 #include "cad_tessellation_modeler.h"
