+++ kratos/python/add_geometrical_utilities_to_python.cpp
@@ -38,7 +38,7 @@
 #include "utilities/iso_printer.h"
 #include "utilities/interval_utility.h"
 #include "utilities/convect_particles_utilities.h"
-#include "utilities/delaunator_utilities.h"
+#include "utilities/tessellation_utilities/delaunator_utilities.h"
 #include "utilities/mls_shape_functions_utility.h"
 
 namespace Kratos {
