+++ kratos/tests/cpp_tests/utilities/test_delaunator_utilities.cpp
@@ -14,7 +14,7 @@
 // Project includes
 #include "testing/testing.h"
 #include "containers/model.h"
-#include "utilities/delaunator_utilities.h"
+#include "utilities/tessellation_utilities/delaunator_utilities.h"
 
 namespace Kratos {
 namespace Testing {
